feat: Update minima theme and skin
Updated to use minima 3.x from GitHub which is unreleased. It has been in development for a few years so I am trying this to see if it is stable enough for my basic use. With that in place it allows me to easily change the skin so for now I have updated it to the dark skin.